Module Trace_fuchsia.Exporter

An exporter, takes buffers with fuchsia events, and writes them somewhere

type t = {

write_bufs : Buf.t Stdlib.Queue.t -> unit;
(*
Takes buffers and writes them somewhere. The buffers are only valid during this call and must not be stored. The queue must not be modified.
*)

flush : unit -> unit;
(*
Force write
*)

close : unit -> unit;
(*
Close underlying resources
*)

}

An exporter, takes buffers and writes them somewhere. This should be thread-safe if used in a threaded environment.

val of_out_channel : close_channel:bool -> out_channel -> t

Export to the channel

parameter close_channel
if true, closing the exporter will close the channel

val of_buffer : Stdlib.Buffer.t -> t