Module Trace_fuchsia.Writer

Write fuchsia events into buffers.

Reference: https://fuchsia.dev/fuchsia-src/reference/tracing/trace-format

module Util : sig ... end

type user_data = Trace_core.user_data

type arg =

| A_bool of bool

| A_float of float

| A_int of int

| A_none

| A_string of string

| A_kid of int64

val arg_of_user_data : user_data -> arg

val args_of_user_data : (string * user_data) list -> (string * arg) list

module I64 : sig ... end

module Str_ref : sig ... end

val truncate_string : string -> string

truncate_string s truncates s to the maximum length allowed for strings. If s is already short enough, no allocation is done.

module Thread_ref : sig ... end

module Metadata : sig ... end

record type = 0

module Argument : sig ... end

module Arguments : sig ... end

module Thread_record : sig ... end

record type = 3

module Event : sig ... end

record type = 4

module Kernel_object : sig ... end

record type = 7