Module Trace_fuchsia_write

Write fuchsia events into buffers.

Reference: https://fuchsia.dev/fuchsia-src/reference/tracing/trace-format

module Util : sig ... end

module Buf : sig ... end

module Output : sig ... end

module Buf_pool : sig ... end

type user_data = Trace_core.user_data

module I64 : sig ... end

module Str_ref : sig ... end

val truncate_string : string -> string

truncate_string s truncates s to the maximum length allowed for strings. If s is already short enough, no allocation is done.

module Thread_ref : sig ... end

module Metadata : sig ... end

record type = 0

module Argument : sig ... end

module Arguments : sig ... end

module Thread_record : sig ... end

record type = 3

module Event : sig ... end

record type = 4

module Kernel_object : sig ... end

record type = 7